+++ game/content/ghplots/peopleplots.py
@@ -40,3 +40,29 @@ class RandomExpertNPC(Plot):
 
     def _is_good_scene(self,nart,candidate):
         return isinstance(candidate,pbge.scenes.Scene) and gears.tags.SCENE_PUBLIC in candidate.attributes
+
+
+#  ***************************
+#  ***   ADD_FAITHWORKER   ***
+#  ***************************
+# Add a faithworker of some type.
+
+class RandomFaithworker(Plot):
+    LABEL = "ADD_FAITHWORKER"
+    JOB_NAMES = ("Priest", "Monk", "Warrior Monk", "Neodruid", "Technoshaman")
+
+    def custom_init(self, nart):
+        npc = gears.selector.random_character(
+            job = gears.jobs.ALL_JOBS[random.choice(self.JOB_NAMES)],
+            age=random.randint(50,90),
+            local_tags=tuple(self.elements["METROSCENE"].attributes)
+        )
+        self.seek_element(nart, "LOCALE", self._is_best_scene, scope=self.elements["METROSCENE"], backup_seek_func=self._is_good_scene)
+        self.register_element("NPC", npc, dident="LOCALE")
+        return True
+
+    def _is_best_scene(self,nart,candidate):
+        return isinstance(candidate, gears.GearHeadScene) and gears.tags.SCENE_PUBLIC in candidate.attributes and candidate.attributes.intersection({gears.tags.SCENE_CULTURE, gears.tags.SCENE_TEMPLE})
+
+    def _is_good_scene(self,nart,candidate):
+        return isinstance(candidate, gears.GearHeadScene) and gears.tags.SCENE_PUBLIC in candidate.attributes